[fix] #48 reduce DTO
