--- src/main/java/com/telepigeon/server/oauth/dto/KakaoUserProfile.java
@@ -1,9 +1,0 @@
-package com.telepigeon.server.oauth.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserProfile (String nickname){
-
-}
